feat: 兼容設定會在更新後同步執行，也會 AS 排程執行

+++ inc/classes/Compatibility/Compatibility.php
@@ -19,13 +19,34 @@ use J7\PowerCourse\AbstractTable;
 final class Compatibility {
 	use \J7\WpUtils\Traits\SingletonTrait;
 
+	const AS_COMPATIBILITY_ACTION = 'pc_compatibility_action_scheduler';
+
 	/** Constructor */
 	public function __construct() {
+		$scheduled_version = \get_option('pc_compatibility_action_scheduled');
+		if ($scheduled_version === Plugin::$version) {
+			return;
+		}
+
 		\delete_option('pc_compatibility_action_scheduled');
 
 		ApiOptimize::instance();
 		// 升級成功後執行
 		\add_action( 'upgrader_process_complete', [ __CLASS__, 'compatibility' ]);
+
+		// 排程只執行一次的兼容設定
+		\add_action( 'init', [ __CLASS__, 'compatibility_action_scheduler' ] );
+		\add_action( self::AS_COMPATIBILITY_ACTION, [ __CLASS__, 'compatibility' ]);
+	}
+
+
+	/**
+	 * 排程只執行一次的兼容設定
+	 *
+	 * @return void
+	 */
+	public static function compatibility_action_scheduler(): void {
+		\as_enqueue_async_action( self::AS_COMPATIBILITY_ACTION, [] );
 	}
 
 
@@ -69,6 +90,9 @@ final class Compatibility {
 		/**
 		 * ============== END 相容性代碼 ==============
 		 */
+
+		// ❗不要刪除此行，註記已經執行過相容設定
+		\update_option('pc_compatibility_action_scheduled', Plugin::$version);
 
 		\J7\WpUtils\Classes\WC::log(Plugin::$version, '已執行兼容性設定');
 	}
